Crate static_vector

Static Vector

A no-std, stack-allocated vector with fixed capacity and dynamic length.

Vec stores elements on the stack using a fixed-size array without heap allocations.

Aims to be suitable for low-level projects and to have an API as safe and explicit as possible. The goal is to allocate only when needed. When first constructed, the vector will not allocate.

Note: It’s a learning project, so there are no guarantees.

Features

No heap allocation (#![no_std] compatible)

Supports iteration, mutable access, clearing, resizing

Compile-time enforced capacity

Requirements

T: Clone for insertion: Vec::push()

T: Default only if Vec::set_len() is used

Complexity

All operations are O(1) except:

Method	Time Complexity	Space Complexity

clear	O(current length)	O(1)

set_len	O(new length - current length)	O(new length - current length)

Example

use static_vector::Vec;

let mut vec = Vec::<i32, 3>::new();

vec.push(&4).unwrap();
vec.push(&5).unwrap();
vec.push(&6).unwrap();
assert_eq!(vec.len(), 3);
assert_eq!(vec.first(), Some(&4));

let sum_of_even_numbers = vec.iter().filter(|n| *n % 2 == 0).sum::<i32>();
assert_eq!(sum_of_even_numbers, 10);

vec.push(&2).unwrap_err();
assert_eq!(vec.len(), 3);

match vec.set_len(1) {
    Ok(()) => assert_eq!(vec.len(), 1),
    Err(err) => eprintln!("{:?}", err),
}

vec.clear();
assert!(vec.is_empty());

Development on Linux

See Makefile.

Structs

CapacityExceededError

Attempted to push to a full vector

Iter

Immutable iterator over a Vec.

IterMut

Mutable iterator over a Vec.

LengthTooLargeError

Attempted to resize the vector to a length greater than its fixed capacity.

Vec

A stack-allocated vector with fixed capacity and dynamic length.