static_vector

Struct Vec

pub struct Vec<T, const CAPACITY: usize> { /* private fields */ }

A stack-allocated vector with fixed capacity and dynamic length.

Implementations

impl<T, const CAPACITY: usize> Vec<T, CAPACITY>

pub const fn new() -> Self

Creates a new empty Vec with maximum CAPACITY elements of type T.

Panics

Panics if CAPACITY == 0. Zero-capacity vectors are not supported.

Example

use static_vector::Vec;

let mut vec = Vec::<i32, 20>::new();
vec.as_mut_slice().fill(0);
// now vec has length 20 and all elements are zero

pub const fn capacity(&self) -> usize

Returns the maximum number of elements the vector can contain.

Example

use static_vector::Vec;

let vec = Vec::<i32, 10>::new();
const SOME_LIMIT: usize = 5;

if vec.len() < vec.capacity() - SOME_LIMIT {
    // do something
}

pub const fn len(&self) -> usize

Returns the number of elements the vector currently contains.

Example

use static_vector::Vec;

let mut vec = Vec::<i32, 10>::new();
const SOME_LIMIT: usize = 5;

if vec.len() < SOME_LIMIT {
    // do something
}

pub const fn is_empty(&self) -> bool

Returns whether the vector has no elements.

Example

use static_vector::Vec;

let mut vec = Vec::<i32, 10>::new();

if vec.is_empty() {
    // do something
}

pub const fn is_full(&self) -> bool

Returns whether the vector is at maximum capacity.

Example

use static_vector::Vec;

let mut vec = Vec::<i32, 20>::new();

if vec.is_full() {
    // cannot push elements anymore
}

pub fn push(&mut self, value: T) -> Result<(), CapacityError>

Adds the given value to the end of the vector.

Errors

Returns CapacityError if the vector is already at full capacity.

Example

use static_vector::{CapacityError, Vec};

#[derive(Debug)]
enum AppError {
    VectorCapacityError(CapacityError),
}

fn my_fn(vec: &mut Vec<i32, 2>) -> Result<(), AppError> {
    vec.push(1).map_err(AppError::VectorCapacityError)?;
    vec.push(1).map_err(AppError::VectorCapacityError)?;

    // third push will fail because vector capacity is 2
    vec.push(3).map_err(AppError::VectorCapacityError)?;

    // other operations that could return errors
    Ok(())
}

fn main() -> Result<(), AppError> {
    let mut vec = Vec::<i32, 2>::new();

    if let Err(err) = my_fn(&mut vec) {
        match err {
            AppError::VectorCapacityError(_) => {
                // handle case
            },
        }
    }

    Ok(())
}

pub fn clear(&mut self)

Removes all elements. Size will be zero.

Example

use static_vector::Vec;

let mut vec = Vec::<i32, 20>::new();

// add some elements
vec.clear();
// elements will be gone

pub fn set_len(&mut self, new_length: usize) -> Result<(), CapacityError>
where T: Default,

Resizes the vector to the new_length.

Requirements

T must implement Default because new elements are created with T::default() when increasing the length.

Errors

Returns CapacityError if new_length exceeds the vector’s fixed capacity.

Example

use static_vector::Vec;

#[derive(Debug)]
enum AppError {
    MyFnError,
}

fn my_fn(vec: &mut Vec<i32, 200>) -> Result<(), AppError> {
    vec.set_len(100).map_err(|_| AppError::MyFnError)?;

    // other operations that could return errors

    Ok(())
}

fn main() -> Result<(), AppError> {
    let mut vec = Vec::<i32, 200>::new();

    if let Err(err) = my_fn(&mut vec) {
        match err {
            AppError::MyFnError => {
                // handle case
            },
        }
    }

    Ok(())
}

pub const fn first(&self) -> Option<&T>

Returns a reference to the first element in the vector, or None if the vector is empty.

Example

use static_vector::Vec;

let mut vec = Vec::<i32, 20>::new();

match vec.first() {
    Some(num) => {
        let _ = num;
    },
    None => {
        // no first element
    },
}

pub const fn first_mut(&mut self) -> Option<&mut T>

Returns a mutable reference to the first element in the vector, or None if the vector is empty.

Example

use static_vector::Vec;

let mut vec = Vec::<i32, 20>::new();

if let Some(num) = vec.first_mut() {
    *num = 1;
    let _ = num;
}

pub const fn last(&self) -> Option<&T>

Returns a reference to the last element in the vector, or None if the vector is empty.

Example

use static_vector::Vec;

let mut vec = Vec::<i32, 30>::new();

if let Some(num) = vec.last() {
    let _ = num;
    // do something with the last element
}

pub const fn last_mut(&mut self) -> Option<&mut T>

Returns a mutable reference to the last element in the vector, or None if the vector is empty.

Example

use static_vector::Vec;

let mut vec = Vec::<i32, 20>::new();

if let Some(num) = vec.last_mut() {
    *num = 1;
    let _ = num;
}

pub const fn get(&self, index: usize) -> Option<&T>

Returns a reference to the element at the specified index, or None if out of bounds.

Example

use static_vector::Vec;

let mut vec = Vec::<i32, 20>::new();

match vec.get(22) {
    Some(num) => {
        let _ = num;
    },
    None => {
        // element with index 22 does not exist
    },
}

pub const fn get_mut(&mut self, index: usize) -> Option<&mut T>

Returns a mutable reference to the element at the specified index, or None if out of bounds.

Example

use static_vector::Vec;

let mut vec = Vec::<i32, 20>::new();

if vec.push(1).is_ok() {
    *vec.get_mut(0).unwrap() = 5;
}

pub const fn pop(&mut self) -> Option<T>

Returns (and removes) the last element from the vector, or None if the vector is empty.

Example

use static_vector::Vec;

let mut vec = Vec::<i32, 20>::new();

// fill vector with number from 1 to 10
vec.set_len(10).unwrap(); // can unwrap because 10 < 20
vec.as_mut_slice().fill_with({
    let mut n = 0;
    move || {
        n = n + 1;
        n
    }
});

// print in reverse order while removing from
while let Some(num) = vec.pop() {
    print!("{num} ");
}

// prints: 10 9 8 7 6 5 4 3 2 1
// the vector is now empty

pub fn pop_if<F: FnOnce(&T) -> bool>(&mut self, predicate: F) -> Option<T>

Returns (and removes) the last element from the vector if the predicate returns true, or None if the vector is empty or the predicate returns false.

Example

Similar to Vec::pop(), but needs a predicate

use static_vector::Vec;

let mut vec = Vec::<i32, 20>::new();

// assuming vector has elements: 10 9 8 7 6 5 4 3 2 1
if let Some(num) = vec.pop_if(|n| *n % 2 == 0) {
    println!("{num}");
    // prints: 10
    // the vector has remained with: 1 2 3 4 5 6 7 8 9
}

pub const fn iter(&self) -> Iter<'_, T>

Returns an iterator over immutable references to the elements in the vector.

Example

use static_vector::Vec;

let mut vec = Vec::<i32, 20>::new();

vec.iter().map(|n| *n * 2).sum::<i32>();

pub const fn iter_mut(&mut self) -> IterMut<'_, T>

Returns an iterator over mutable references to the elements in the vector.

Example

use static_vector::Vec;

let mut vec = Vec::<i32, 20>::new();

for num in vec.iter_mut() {
    *num *= 2;
}

pub const fn as_slice(&self) -> &[T]

Returns a slice of the entire vector.

Example

use static_vector::Vec;

fn count_ones(numbers: &[i32]) -> usize {
    numbers.iter().filter(|n| *n == &1).count()
}

let vec = Vec::<i32, 1000>::new();

// push numbers into vector
     
if vec.as_slice().binary_search(&1).is_ok() {
    // found it
}

let ones = count_ones(vec.as_slice());
if ones > 0 {
    // ...
}

pub const fn as_mut_slice(&mut self) -> &mut [T]

Returns a mutable slice of the entire vector.

Example

use static_vector::Vec;

let mut vec = Vec::<i32, 10>::new();

if vec.set_len(5).is_ok() {
    vec.as_mut_slice().fill(1);
} else {
    // handle error
}

pub fn extend_from_slice(&mut self, slice: &[T]) -> Result<(), CapacityError>
where T: Clone,

Inserts elements of given slice at the end of the vector.

Errors

Returns CapacityError if adding elements of given slice would result in vector exceeding its capacity.

Example

use static_vector::{CapacityError, Vec};

#[derive(Debug)]
enum AppError {
    VectorCapacityError(CapacityError),
}

fn my_fn(src: &[i32], vec: &mut Vec<i32, 2>) -> Result<(), AppError> {
    vec.extend_from_slice(src).map_err(AppError::VectorCapacityError)?;

    // other operations that could return errors
    Ok(())
}

fn main() -> Result<(), AppError> {
    let src = [1, 2, 3];
    let mut vec = Vec::<i32, 2>::new();

    if let Err(err) = my_fn(&src, &mut vec) {
        match err {
            AppError::VectorCapacityError(_) => {
                // handle case
            },
        }
    }

    Ok(())
}

pub fn append<const OTHER_CAPACITY: usize>( &mut self, other: &mut Vec<T, OTHER_CAPACITY>, ) -> Result<(), CapacityError>
where T: Clone,

Moves elements of other vector at the end of the current vector. other will be empty.

Errors

Returns CapacityError if adding elements from other would result in current vector exceeding its capacity.

Example

use static_vector::Vec;

let mut vec = Vec::<i32, 10>::new();
vec.push(1).unwrap();
vec.push(2).unwrap();
vec.push(3).unwrap();

let mut other = Vec::<i32, 20>::new();
other.push(4).unwrap();
other.push(5).unwrap();
other.push(6).unwrap();

vec.append(&mut other).unwrap();
assert_eq!(vec.as_slice(), [1, 2, 3, 4, 5, 6]);
assert_eq!(other.as_slice(), []);